Centralize more column resolution rules
